Redirect to static page after search
Previously, searching for a gene would result in a url like
'/search/genesymbol' with the symbol being POSTed to the web server.
Now the search page will redirect to '/info/gene/<sym>' to display
the page for the gene or an error message. This is better because
it lets a user link straight to a page for later reference.